refactor(igx-ts): move `relativePath` + `formatOutput` to Util file
