+++ lib/project-utility/TypeScriptUtils.ts
@@ -1,7 +1,6 @@
 // tslint:disable-next-line:no-implicit-dependencies
 import { Tree } from "@angular-devkit/schematics/src/tree/interface";
 import * as fs from "fs";
-import * as path from "path";
 import * as ts from "typescript";
 import { Util } from "../Util";
 
@@ -95,39 +94,6 @@ export class TypeScriptUtils {
 	//#endregion Node manipulation utils
 
 	//#region Utility functions
-
-	/**
-	 * Generate relative path from target file to another
-	 * Adds "./" to avoid node module resolution conflicts
-	 * @param targetPath Target file (root path)
-	 * @param filePath File to generate relative path to
-	 * @param posix Require path in posix style (/-separated)
-	 * @param removeExt Strip file extension
-	 */
-	public static relativePath(targetPath: string, filePath: string, posix: boolean, removeExt = true): string {
-		if (!targetPath.endsWith(path.win32.sep) && !targetPath.endsWith(path.posix.sep)) {
-			// path.relative splits by fragments, must be dirname w/ trailing to work both down and up
-			targetPath = path.win32.dirname(targetPath) + path.sep;
-		}
-
-		// use win32 api as it handles both formats
-		let relativePath: string = path.win32.relative(targetPath, filePath);
-
-		if (removeExt) {
-			relativePath = relativePath.replace(path.win32.extname(relativePath), "");
-		}
-
-		if (posix) {
-			relativePath = path.posix.join(...relativePath.split(path.win32.sep));
-			relativePath = relativePath.startsWith(".") ? relativePath : "./" + relativePath;
-		} else {
-			relativePath = path.win32.join(...relativePath.split(path.win32.sep));
-			relativePath = relativePath.startsWith(".") ? relativePath : ".\\" + relativePath;
-		}
-
-		return relativePath;
-	}
-
 	/**
 	 * Returns an source file, adds new line placeholders as the TS parser won't add `SyntaxKind.NewLineTrivia` to the AST.
 	 * @param filePath Path of file to read
