+++ spec/unit/ts-transform/TypeScriptUtils-spec.ts
@@ -5,72 +5,6 @@ import { FsFileSystem, TypeScriptUtils } from "../../../lib/project-utility/Type
 describe("Unit - TypeScriptUtils", () => {
 	const fileSystem = new FsFileSystem();
 
-	describe("Relative paths", () => {
-		it("Creates correct relative path for child structure", async done => {
-
-			// default use, shared root
-			expect(TypeScriptUtils.relativePath(
-				"C:\\src\\app\\app-routing.module.ts",
-				"C:\\src\\app\\carousel\\carousel.component.ts", true, true
-			))
-			.toBe("./carousel/carousel.component", "Shared Win root, file to file => posix ");
-			expect(TypeScriptUtils.relativePath(
-				"/home/app/app.module.ts",
-				"/home/app/grid/grid.component.ts", true, true
-			))
-			.toBe("./grid/grid.component", "Shared posix root, file to file => posix ");
-
-			expect(TypeScriptUtils.relativePath(
-				"C:\\home\\app-routing.module.ts",
-				"C:/home/app/grid/grid.component.ts", true, false
-			))
-			.toBe("./app/grid/grid.component.ts", "Win style to posix file, keep ext => posix");
-			done();
-		});
-
-		it("Creates correct relative path for parent dir", async done => {
-			// going up levels
-			expect(TypeScriptUtils.relativePath(
-				"C:\\common\\folder1\\folder2\\folder3\\file.ts",
-				"C:\\common\\dir1\\dir2\\target.ts", false
-			))
-			.toBe("..\\..\\..\\dir1\\dir2\\target", "Win style, file to file => win");
-
-			expect(TypeScriptUtils.relativePath(
-				"C:\\Work\\git\\Ignite-UI-CLI\\spec\\unit\\ts-transforms\\TypeScriptUtils-spec.ts",
-				"C:\\Work\\git\\Ignite-UI-CLI\\lib\\project-utility\\TypeScriptUtils.ts", true
-			))
-			// same as top level import above
-			.toBe("../../../lib/project-utility/TypeScriptUtils", "Win style, file to file => posix");
-
-			expect(TypeScriptUtils.relativePath(
-				"C:\\common\\folder1\\folder2\\folder3\\",
-				"C:\\common\\dir1\\dir2\\target.ts", true
-			))
-			.toBe("../../../dir1/dir2/target", "Win style, folder to file => posix");
-
-			expect(TypeScriptUtils.relativePath(
-				"C:\\common\\folder1\\folder2\\folder3\\",
-				"C:\\common\\dir1\\dir2\\target", true
-			))
-			.toBe("../../../dir1/dir2/target", "Win style, folder to file w/o ext => posix");
-
-			expect(TypeScriptUtils.relativePath(
-				"C:/common/folder1/folder2/folder3/",
-				"C:\\common\\dir1\\dir2\\target.ts", true
-			))
-			.toBe("../../../dir1/dir2/target", "posix folder to Win style file => posix");
-
-			expect(TypeScriptUtils.relativePath(
-				"C:/common/folder1/folder2/folder3/",
-				"C:\\common\\dir1\\dir2\\target.ts", true, false
-			))
-			.toBe("../../../dir1/dir2/target.ts", "posix folder to Win style file, keep ext => posix");
-
-			done();
-		});
-	});
-
 	describe("File read/write", () => {
 		const newLines = {
 			CRLF: "\r\n",
